requirements.txt
click==8.1.7
pydantic>=2.0.0,<3.0.0
openai>=1.0.0
python-dotenv==1.0.1
